+++ server/server.js
@@ -11,6 +11,7 @@ const source = 'public';
 
 const serveAppFrom = path.resolve() + '/' + source;
 const serveDocsFrom = path.resolve() + '/docs/';
+const serveCreateFrom = path.resolve() + '/create/';
 
 const app = express();
 
@@ -45,19 +46,21 @@ app.use('/', function (req, res, next) {
 
 // don't return anything from the api directory
 app.use(function (req, res, next) {
-  if (req.url.indexOf('api') > -1) {
+  if (req.url.indexOf('api') > -1 && (req.url.indexOf('.js') > -1)) {
     res.sendStatus(403);
   } else {
     next();
   }
 });
 
-app.use(express.static(serveAppFrom, {
-
-  fallthrough: true
-}));
+app.use(express.static(serveAppFrom, { fallthrough: true }));
 
 app.use(express.static(serveDocsFrom, { fallthrough: true }));
+
+/* moonsault create app */
+app.use(express.static(serveCreateFrom, { fallthrough: true }));
+const createServices = require(`${serveCreateFrom}/create/api/services.js`);
+createServices.start(app);
 
 app.get('/', (req, res, next) => {
   next();